elif -> if for dict check in modify_dependencies
This is a subtle change. The idea is that if you replace the obj
with a dict, you can process the dict further. In the case where
you're doing a deep to_dict, this just means that you recurse into
the next dict before assigning the unmodified one to your current
location, and (I'm pretty sure) it makes no other difference.

The value is that it enables cases where you might want to have
a tokenizable in the dict that replaces the tokenizable. Example:

def get_all_gufe_objs(obj):
    results = set()
    def modifier(o):
        results.add(o)
        return o.to_shallow_dict()

    _ = modify_dependencies(obj.to_shallow_dict(), modifier, is_gufe_obj)
    return results