Added normalization of dicts, lists, sets, etc; dag_scratch and unit_scratch
We needed normalization similar to how `dask.base` implements it in
order to ensure that data structures of nested lists, dicts, and tuples
are tokenized in a deterministic way. We utilize here a subset of the
data structures `dask.base` supports, and can add more as needed or
adopt `dask.base` wholesale later if really needed.

Also added the concepts of `dag_scratch` and `unit_scratch` to give
protocol developers handles for whre to put files in their protocols,
with certain guarantees in place. These should be used by the Scheduler
that executes `ProtocolDAG`s with appropriately-configured filesystem
locations that are optimal for each.