Reworked some of the ugly structure of ProtocolDAG and ProtocolDAGResult
This was motivated in part by removing inclusion of the networkx graph
representation from `ProtocolDAGResult._to_dict` for serializability.
Also wanted to remove reliance on `networkx` graph for `'result'`
references, and instead rely on mappings for this.

We can of course change these internals in the future if needed.
Now there is greater API parity between ProtocolDAG and
ProtocolDAGResult.