Fixed decoding issue with `modify_dependencies`
We need two distinct modes for `modify_dependencies` to work as
intended: one for encoding and one for decoding. This impacts when we
convert a `GufeTokenizable` into or out of its dict/keyed form. Prior to
this change we were decoding objects before traversing down through them
for other `GufeTokenizable`s.

Also included here is the `dependencies` property for `ProtocolUnit`s
and `ProtocolUnitResult`s. This simplified the graph builder for
`ProtocolDAG` and `ProtocolDAGResult`.