Removed concept of pure from DAG system
With the addition of `dag_scratch`, it's not clear that there is any
benefit to deduplication of `ProtocolUnit`s, since which `dag_scratch` a
unit present among multiple DAGs would use is ambiguous.

To simplify our approach, decided to remove `pure` as a concept for DAG
system, and just make `ProtocolUnit` keys based on UUID assigned on
creation. These objects include their key in their `to_dict` form, and
inject this on creation via `from_dict`. This allows for clean
roundtripping without breaking the model.