DAGs now constructed using ProtocolUnit inputs; inputs can be mapped into keyword arguments of `ProtocolUnit._execute` (#36)
* Working on making dags more friendly for protocol authors

About to do some culling on the graph model; no need for graph to hold
inputs on edges if we've already tokenized inputs on protocolunits

* Unit tests working again, with some additions.

* Now use keys instead of the tokens for protocolunit identification

key = <classname>-<token>

This may help to avoid collisions, and includes some visual clue about
what the object being referred to actually is better than just the
token. We use this as the name of the object if no name is specified.

* Added dask-core to conda-forge dependencoes; annotation fixes

* Small doc additions

* Working to make `ProtocolUnit`s `GufeTokenizable`s

* Hacking away at making `ProtocolUnit`s work as `GufeTokenizable`s

* Tests passing with `GufeTokenizable`s for DAG system

A bit slow, however, due probably to tokenization all the way down for
the last `ProtocolUnit` in a DAG. Considering approaches we might take
to improve performance here, such as tokenization of results?

* Deepcopy no longer necessary in `modify_dependencies`

Slows things down to a crawl for nested objects, and likely eats memory
too. With the way `modify_dependencies` is written, there should be no
need for any deep copies now.

* Added normalization of dicts, lists, sets, etc; dag_scratch and unit_scratch

We needed normalization similar to how `dask.base` implements it in
order to ensure that data structures of nested lists, dicts, and tuples
are tokenized in a deterministic way. We utilize here a subset of the
data structures `dask.base` supports, and can add more as needed or
adopt `dask.base` wholesale later if really needed.

Also added the concepts of `dag_scratch` and `unit_scratch` to give
protocol developers handles for whre to put files in their protocols,
with certain guarantees in place. These should be used by the Scheduler
that executes `ProtocolDAG`s with appropriately-configured filesystem
locations that are optimal for each.

* Switched to use of Context dataclass for passing scratch dirs to
`ProtocolUnit._execute`.

* Drop dask-core from environment

* Rename tokenize.py to tokenization.py

Avoids a collision with standard lib tokenize module; not sure if this
is causing our weird coverage issues

* Remove unnecessary bits from `ProtocolUnit`

* Remove developer install from CI

Attempting to understand coverage issue.
Appears related to existence of `gufe.egg-info` being linked to via
development install. Installation not necessary to run tests.

* Annotation fix

* Reworked some of the ugly structure of ProtocolDAG and ProtocolDAGResult

This was motivated in part by removing inclusion of the networkx graph
representation from `ProtocolDAGResult._to_dict` for serializability.
Also wanted to remove reliance on `networkx` graph for `'result'`
references, and instead rely on mappings for this.

We can of course change these internals in the future if needed.
Now there is greater API parity between ProtocolDAG and
ProtocolDAGResult.

* Restore development install in CI; run coverage outside pytest

* Accommodate our exposure of fixtures as entrypoints

We want to avoid the issue where modules are imported before coverage
starts running due to our exposure of some fixtures as entrypoints.

See: https://github.com/pytest-dev/pytest-cov/issues/455#issuecomment-1026836811

* Added note to CI

* Revert pytest usage in CI; remove fixture entrypoint

Will add fixture entrypoint in a separate PR if we actually need it.

* Protocol and ProtocolResult now `GufeTokenizable`s

* Switched default equality for GufeTokenizable to use keyed_dict

By doing this, we avoid the issue of equality between e.g. NaNs from
objects a GufeTokenizable is dependent on. If two GufeTokenizables give
identical keyed dicts (their dependencies are the same objects), then
they are considered equal.

This should be very reliable given how we've constructed the
TOKENIZABLE_REGISTRY, I think.

* Moved execute method out of ProtocolDAG as standalone function

This is intended mainly for developers, not really for users, so moving
it out of ProtocolDAG API.

* Fix broken transformation tests

* Docstrings, type hints

* Fixed mypy complaints

* Fixed decoding issue with `modify_dependencies`

We need two distinct modes for `modify_dependencies` to work as
intended: one for encoding and one for decoding. This impacts when we
convert a `GufeTokenizable` into or out of its dict/keyed form. Prior to
this change we were decoding objects before traversing down through them
for other `GufeTokenizable`s.

Also included here is the `dependencies` property for `ProtocolUnit`s
and `ProtocolUnitResult`s. This simplified the graph builder for
`ProtocolDAG` and `ProtocolDAGResult`.

* Add dependency tests for ProtocolDAGs; change `GufeTokenizable.__eq__`

We don't want use of e.g. `in` to throw an exception if two
`GufeTokenizable`s are not the same class; we just want them to give
`False`.

* Removed concept of pure from DAG system

With the addition of `dag_scratch`, it's not clear that there is any
benefit to deduplication of `ProtocolUnit`s, since which `dag_scratch` a
unit present among multiple DAGs would use is ambiguous.

To simplify our approach, decided to remove `pure` as a concept for DAG
system, and just make `ProtocolUnit` keys based on UUID assigned on
creation. These objects include their key in their `to_dict` form, and
inject this on creation via `from_dict`. This allows for clean
roundtripping without breaking the model.

* Updated docstrings for major components of DAG system

* Mypy now happy

* Added Context to protocol module __init__

* Small changes from meeting with @ijpulido

* Remove _key from init on ProtocolUnit, test fix

* Removed ProtocolDAGFailure; made ProtocolUnitFailure.exception ser/deser friendly

* dag_scratch -> shared; unit_scratch -> scratch

* Added docstrings to Protocol overridables.

Still need docstrings for ProtocolResult overridables, but will need
feedback from first protocols on what makes most sense for those.