Stuff for loading PDB via URL
I'm not sure that I'm getting the PDBs from the right place, but this
commit gives testing tools for loading a PDB from a URL (loading it
into a StringIO).

It is a little weird in that it returns local files as string filename
and URLs as filelike StringIO objects; it might be better to have both
return filelike and have a separate test for the ability to load as
filename or filelike.