Speed up protein test loading
This essentially makes our predefined PDB loaders session-scoped. This
cuts out a lot of network activity and should give us as notable speedup
when running the full test suite.

In principle, it would probably be better to structure everything to use
the `request` fixture, but this is a smaller code change for essentially
the same result.