Proposed fix to dag ordering (#315)
* Propose fix to dag ordering

`lexicographical_topological_sort` creates a unique ordering of `ProtocolUnits` when the `GufeKey`s of those units are used for comparison. This adds stability to the subsequent `GufeKey`s of DAGs containing these units.

* Modified test_execute_DAG_retries

Using `GufeKeys` to further sort DAGs with
`lexicographical_topological_sort`, the execution order of
`ProtocolUnits` for this particular `Protocol` differed from the
original implementation using `topological_sort`. This new sorting
placed the guaranteed-to-fail `ProtocolUnit` later, allowing `ok`
results to accumulate. This change reflects the new ordering,
resulting in 26 total results instead of 5, now accounting for the 21
`SimulationUnits` that complete before `BrokenSimulationUnit` fails.

---------

Co-authored-by: Ian Kenney <ianmichaelkenney@gmail.com>