Pip check doesn't work with conda and pip
We get these errors with pip check:

mdtraj 1.10.3 requires netcdf4, which is not installed.
sphinx 7.1.1 has requirement docutils<0.21,>=0.18.1, but you have docutils 0.21.2.

netcdf4 must be an optional dep with the conda-forge package, so this
will always fail. When we mix conda and pip to install things, we
shouldn't expect pip check to work anyway