Use correct default shell (#489)
* Use a default shell that shows errors

* pretty sure newer rdkit has fixed the typing issue

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Pip check doesn't work with conda and pip

We get these errors with pip check:

mdtraj 1.10.3 requires netcdf4, which is not installed.
sphinx 7.1.1 has requirement docutils<0.21,>=0.18.1, but you have docutils 0.21.2.

netcdf4 must be an optional dep with the conda-forge package, so this
will always fail. When we mix conda and pip to install things, we
shouldn't expect pip check to work anyway

* remove -a from micromamba call

* pin to working rdkit version to fix typing

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Alyssa Travitz <31974495+atravitz@users.noreply.github.com>