Add `Protocol` validation method (#502)
* Add _validate and validate methods to `Protocol`

* Implement `Protocol` validation pass-through in `Transformations`

`Transformation` and `NonTransformation` constructors now allow the
user to optionally specify the `validate` kwarg. If `True` the
constructor will provide its parameters as inputs to the `Protocol`
`validate` method (less `extends`) in the same way `create` would.

The default behavior is to set this as `False`.

* Make _validate optional, but provide a surface-level default validation

Authors will not be required to implement the `_validate` method for
`Protocol` subclasses. The validate method now runs the
`_default_validate` method, which effectively just checks types as of
this commit, and will then attempt to run an author-defined `_validate`
method, excepting the `NotImplementedError` in the event that the author
did not implement it.

This is the default implementation in `Protocol`.

---------

Co-authored-by: David Dotson <dotsdl@gmail.com>