Remove 181l_openmmClean.pdb
I am not sure exactly why we have 181l_openmmClean.pdb and 181l.pdb. As
far as I can tell, the "clean" version has been through openMM and the
181l.pdb was "raw" from the openfe benchmark repo. The issue with
keeping the "raw" version is that we need the charges in the correct
location. So now the "clean" and "raw" version are the same. Don't
worry, we are still testing "raw" versions of pdbs from the openfe
benchmark repo.