Minimal changes to support openmm 8.2 and 8.1.2 (#539)
* the smallest changes really needed

* test openmm 8.2 and 8.1.2

* put charges where they belong

* support openmm 8.1.2 and 8.2

* update test to match reformted pdb file

* Remove 181l_openmmClean.pdb

I am not sure exactly why we have 181l_openmmClean.pdb and 181l.pdb. As
far as I can tell, the "clean" version has been through openMM and the
181l.pdb was "raw" from the openfe benchmark repo. The issue with
keeping the "raw" version is that we need the charges in the correct
location. So now the "clean" and "raw" version are the same. Don't
worry, we are still testing "raw" versions of pdbs from the openfe
benchmark repo.

* unpin openmm from env file

* add packaging as a dep + sort deps

* skip testing older openmm on tests that will fail, also start to support openmm as a soft dep

* Update gufe/components/proteincomponent.py

* fix in inequality

* Add license file for vendored code

* Added news entry

* test only on openmm 8.2

---------

Co-authored-by: Irfan Alibay <IAlibay@users.noreply.github.com>